==> 01.check_annot_size.ipynb <==
import pandas as pd
import os
import glob

# PARAMETERS
# inputs = {"__09.19_hyper": "hyper_DMRs_09.19", 
#           "__09.19_hypo": "hypo_DMRs_09.19", 
# #         "hyper": "hyper_DMRs_09.26", 
# #          "hypo": "hypo_DMRs_09.26", 
#           "group-loop" : "group_loops.09.19",
#           "subclass-loop" : "subclass_loops.09.19", 
#           "ct-hyper" : "ct-hyper_DMRs.09.26", 
#           "ct-hypo" : "ct-hypo_DMRs.09.26", 
#           "10.02_hyper" : "hyper_DMRs_10.02", 
#           "10.02_hypo": "hypo_DMRs_10.02",
#           "group-hyper" : "group-hyper_10.07", 
#           "group-hypo" : "group-hypo_10.07"}

# inputs = { "group-hyper" : "group-hyper_10.13", 
#           "group-hypo" : "group-hypo_10.13"}

# inputs = {
#     "msn-hypo" : "msn-hypo-12.02", 
#     "msn-hyper" : "msn-hyper-12.02",
# }

inputs = {
    "abc-group" : "abc-group-12.22", 
    "abc-subclass" : "abc-subclass-12.22",
}


BASEDIR = "/anvil/projects/x-mcb130189/rwang22/bican/ldsc/annot"

for key, value in inputs.items():
    # instantite empty df
    ct_snp_count = {}
    # iterate through each cell type 
    INDIR=f'{BASEDIR}/{key}'
    for ct in os.listdir(INDIR):
        CURDIR = f'{INDIR}/{ct}'
        if os.path.isdir(CURDIR):
            # print(ct)
            count = 0
            for chr in glob.glob(f'{CURDIR}/*.annot.gz'):
                df = pd.read_csv(chr, header = 0, compression = 'gzip')
                tmp = df["ANNOT"].sum()
                count = count + tmp
            ct_snp_count[ct] = count
    count_summary = pd.DataFrame.from_dict(ct_snp_count, orient = "index")
    count_summary["90k_Plus"] = count_summary[0] > 90000
    count_summary.to_csv(f'annot_counts/{value}_annot_counts.tsv', sep = "\t")
