==> 05.dmr_signif_associations_disease_loci.ipynb <==
from pybedtools import BedTool
import pandas as pd
import glob
import os 

# PARAMETERS
input_dir = "/anvil/projects/x-mcb130189/Wubin/BG/pseudobulk/DMR/Subclass"
gwas_studies = "/anvil/scratch/x-rwang22/GWAS/signif_neuro_gwas_studies.csv"
genes_bed = "/anvil/projects/x-mcb130189/rwang22/references/hg38/gencode.v48.gene.bed"
chrom_sizes = "/anvil/projects/x-mcb130189/rwang22/references/hg38/hg38.chrom.sizes"

# snp : rsid - rs1950834
# chr14:41605321
# make the snps dataframe 
snp_df = pd.DataFrame({
    "chrom": ["chr14"], 
    "start": [41605320], 
    "end": [41605321]
})

snp = BedTool.from_dataframe(snp_df)

brain_gwas = pd.read_csv(gwas_studies, index_col = 0).rename(columns = {"SNP": "SNPS"})
brain_gwas.head()

brain_gwas["trait"].unique()

gwas_bed = brain_gwas[["CHR", "BP", "SNPS"]].rename(columns = {"BP": "END", "CHR": "CHR_ID"})
gwas_bed = gwas_bed[~gwas_bed["END"].isna()] # variants with NaN position likely could not be mapped to hg38 -> discard 
gwas_bed["CHR_ID"] = gwas_bed["CHR_ID"].apply(lambda x: f"chr{int(x)}") 
gwas_bed["START"] = gwas_bed["END"].apply(lambda x: x - 1) 
gwas_bed["START"] = gwas_bed["START"].astype(int)
gwas_bed["END"] = gwas_bed["END"].astype(int)
gwas_bed = gwas_bed[["CHR_ID", "START", "END", "SNPS"]].drop_duplicates()
gwas_bed 

gwas = BedTool.from_dataframe(gwas_bed).sort()
genes = BedTool(genes_bed).sort()
closest = gwas.closest(genes, d=True).to_dataframe()

closest = closest[["chrom", "start", "end", "name", "thickEnd"]]
closest = BedTool.from_dataframe(closest)
closest.head()

# dmr beds

bed_files = {f: BedTool(f) for f in glob.glob(f'{input_dir}/*/bed/*.dmr.bed')}

df_list = []
for bed in bed_files:
    output = bed_files[bed].intersect(closest, wb = True)
    if len(output) > 1:
        tmp = output.to_dataframe()
        filename = os.path.basename(bed)
        tmp["celltype"] = filename
        df_list.append(tmp)

dmr_overlap = pd.concat(df_list, ignore_index = True)

dmr_overlap = dmr_overlap[["chrom", "start", "end", "thickStart", "thickEnd", "celltype"]].rename(columns = {"chrom": "CHR_ID", 
                                                                                       "start" : "START", 
                                                                                       "end": "END", 
                                                                                       "thickStart": "SNPS", 
                                                                                       "thickEnd" : "Gene"})
dmr_overlap = dmr_overlap.merge(brain_gwas, how = "left", on ="SNPS")[["CHR_ID", "START", "END", "SNPS", "Gene", 
                                                                       "trait", "celltype", "A1", "A2", "P"]]

dmr_overlap.to_csv('dmr_signif_snps_overlap.tsv', sep = "\t", index = False)

# cell type specific bed

# /anvil/projects/x-mcb130189/Wubin/BG/pseudobulk/DMR/Subclass/Neuron/cell_type_specific_bed
cell_type_specific_bed_files = {f: BedTool(f) for f in glob.glob(f'{input_dir}/*/cell_type_specific_bed/*.dmr.bed')}

df_list = []
for bed in cell_type_specific_bed_files:
    output = cell_type_specific_bed_files[bed].intersect(closest, wb = True)
    if len(output) > 1:
        tmp = output.to_dataframe()
        filename = os.path.basename(bed)
        tmp["celltype"] = filename
        df_list.append(tmp)

dmr_overlap = pd.concat(df_list, ignore_index = True)

dmr_overlap = dmr_overlap[["chrom", "start", "end", "thickStart", "thickEnd", "celltype"]].rename(columns = {"chrom": "CHR_ID", 
                                                                                       "start" : "START", 
                                                                                       "end": "END", 
                                                                                       "thickStart": "SNPS", 
                                                                                       "thickEnd" : "Gene"})
dmr_overlap = dmr_overlap.merge(brain_gwas, how = "left", on ="SNPS")[["CHR_ID", "START", "END", "SNPS", "Gene", 
                                                                       "trait", "celltype", "A1", "A2", "P"]]

dmr_overlap.to_csv('ct_specific_dmr_signif_snps_overlap.tsv', sep = "\t", index = False)
