==> 05.top_associations_disease_loci.ipynb <==
from pybedtools import BedTool
import pandas as pd
import glob
import os 

# PARAMETERS
# input_dir = "/anvil/projects/x-mcb130189/Wubin/BG/pseudobulk/DMR/Subclass"
input_dir = "/anvil/scratch/x-rwang22/bican/dmrs/delta_0.2"
gwas_dir = "/anvil/scratch/x-rwang22/GWAS/brain"

# snp : rsid - rs1950834
# chr14:41605321
# make the snps dataframe 
snp_df = pd.DataFrame({
    "chrom": ["chr14"], 
    "start": [41605320], 
    "end": [41605321]
})

snp = BedTool.from_dataframe(snp_df)

# read in all brain-related gwas 
files = glob.glob(f'{gwas_dir}/*.tsv')

dfs = []
for f in files:
    df = pd.read_csv(f, sep = "\t")
    dfs.append(df)

brain_gwas = pd.concat(dfs, ignore_index = True)
select_cols = ["DISEASE/TRAIT", "CHR_ID", "CHR_POS", "MAPPED_GENE", "SNPS"]
brain_gwas = brain_gwas[select_cols].drop_duplicates()
brain_gwas.head()

brain_gwas[brain_gwas["MAPPED_GENE"] == "LRFN5"]

brain_gwas["DISEASE/TRAIT"].unique()

gwas_bed = brain_gwas[["CHR_ID", "CHR_POS", "SNPS"]].rename(columns = {"CHR_POS": "END"})
gwas_bed = gwas_bed[~gwas_bed["END"].isna()] # variants with NaN position likely could not be mapped to hg38 -> discard 
gwas_bed["CHR_ID"] = gwas_bed["CHR_ID"].apply(lambda x: f"chr{x}") 
gwas_bed["START"] = gwas_bed["END"].apply(lambda x: x - 1) 
gwas_bed["START"] = gwas_bed["START"].astype(int)
gwas_bed["END"] = gwas_bed["END"].astype(int)
gwas_bed = gwas_bed[["CHR_ID", "START", "END", "SNPS"]].drop_duplicates()
gwas_bed 

# bed_files = {f: BedTool(f) for f in glob.glob(f'{input_dir}/*/bed/*.dmr.bed')}
bed_files = {f: BedTool(f) for f in glob.glob(f'{input_dir}/*.dmr.bed')}
gwas = BedTool.from_dataframe(gwas_bed)

df_list = []
for bed in bed_files:
    output = bed_files[bed].intersect(gwas, wb = True)
    if len(output) > 1:
        tmp = output.to_dataframe()
        filename = os.path.basename(bed)
        tmp["celltype"] = filename
        df_list.append(tmp)

dmr_overlap = pd.concat(df_list, ignore_index = True)
dmr_overlap = dmr_overlap[["chrom", "start", "end", "thickStart", "celltype"]].rename(columns = {"chrom": "CHR_ID", 
                                                                                       "start" : "START", 
                                                                                       "end": "END", 
                                                                                       "thickStart": "SNPS"})

dmr_overlap = dmr_overlap.merge(brain_gwas, how = "left", on ="SNPS")[["CHR_ID_x", "START", "END", "SNPS", "DISEASE/TRAIT", "celltype", "MAPPED_GENE"]].rename(
    columns = {"CHR_ID_x": "CHR_ID"})
dmr_overlap 

dmr_overlap.to_csv('genetic_variant/dmr_overlap_10.02.tsv', sep = "\t", index = False)
